[breaking] Add support for ISO standards, add JPEG XL ISO/IEC spec (#1192)
To add ISO/IEC standards related to JPEG XL (see #1089), there needs to be a
way to add an entry in browser-specs that has a canonical URL but no actual
nightly URL, because ISO standards are not public (see also #1191).

This update amends the code to allow and create entries without a `nightly`
property when needed (only for ISO standards for now). The code also retrieves
the name of the group that develops an ISO standard.

This is a BREAKING CHANGE because the `nightly` property used to be mandatory.
Projects that expect to find a `nightly.url` property need to be updated to
only look at the root `url` property or to skip the entry altogether.

First ISO spec added to the list is JPEG XL, which will appear as:

```json
{
  "url": "https://www.iso.org/standard/85253.html",
  "seriesComposition": "full",
  "shortname": "iso18181-2",
  "series": {
    "shortname": "iso18181-2",
    "currentSpecification": "iso18181-2",
    "title": "Information technology — JPEG XL image coding system — Part 2: File format",
    "shortTitle": "JPEG XL: File Format"
  },
  "shortTitle": "JPEG XL: File Format",
  "organization": "ISO/IEC",
  "groups": [
    {
      "name": "ISO/IEC JTC 1/SC 29",
      "url": "https://www.iso.org/committee/45316.html"
    }
  ],
  "title": "Information technology — JPEG XL image coding system — Part 2: File format",
  "source": "specref",
  "categories": [
    "browser"
  ],
  "standing": "good"
}
```

Worth noting: the absence of a `nightly` property means that there is no place
to store the status of the spec, which could in theory be "Under development"
or "Published" for ISO specs (there are additional stages in the ISO process
but they are probably not worth capturing in any case). An alternative would be
to have a `nightly.status` property, but that seems clumsy.