Add GIF and JPEG (JFIF) specs
This adds the GIF and JPEG specifications to the list, as requested in #1089.

For GIF, the most canonical URL is the specification published under `w3.org`,
consistent with Specref and Wikipedia.

The notion of group does not mean much for the GIF spec. Also, the link to the
CompuServe website seems a bit wrong, but a URL is needed...

For JPEG, the usual reference is to the interchange format (JFIF). A PDF
version of the 1992 spec exists under `w3.org` at:
https://www.w3.org/Graphics/JPEG/jfif3.pdf

That is the spec that Specref returns for JPEG. That said, JPEG is published
and maintained by ISO/IEC under the name `iso10918-5`:
https://www.iso.org/standard/54989.html

That spec (and name) is used in the list instead of the PDF version because it
provides a more official entry point. If we want to record the public version
from 1992 somewhere, a new "JPEG" or "JPEG-1992" entry could perhaps be
created.