Add OpenGL ES Shading Language and glTF 2.0 (#1238)
This adds 2 additional specifications developed by the Khronos Group, as
requested in #1089, and adjusts the code that was a bit too specific to WebGL
specs and extensions accordingly.

The notion of "Working Group" within the Khronos Group appears somewhat
informal, but the terminology is used in the page that lists Working Group
officers at least: https://www.khronos.org/about/working-group-officers/

Issue #1089 mentions GLSL, and not OpenGL **ES** Shading Language, but the
latter is what WebGL uses, and so should be what browsers implement. The name
`ESSL` is the official name for the ES variant, see:
https://github.com/KhronosGroup/GLSL

That repository only contains extensions and the issue tracker for the GLSL and
ESSL specifications, but not the actual specifications. I did not set the
`nightly.repository` property as a result (but that may be up for debate).

glTF is not added to the "browser" category because browsers do not directly
support glTF today.