Add "browser-specs" CLI and issue workflow to manage lists (#1270)
This adds a new CLI, to be run with `npx browser-specs`. The CLI wraps
`build-diff.js` and eases the insertion of individual specs to the lists
(`specs.json`, `monitor.json` and `ignore.json`). In particular, the command:

- reports changes that would be made to `index.json`;
- runs tests against these changes;
- takes care of linting;
- removes entries from the monitor/ignore list if needed;
- can commit changes and create a pull request out of them.

See inline command help for details.

The goal is to have a command that is both useful for local execution and for
processing issues on GitHub. The CLI reports markdown for that reason.

For a local run, the build command takes a URL as parameter (unless the goal
is to test changes already made to `specs.json`). This should be all that's
needed to add a spec in about 90% of cases. Additional properties can be
specified through the `--json` option.

The build command may also take an issue number as parameter, in which case it
will retrieve the spec information from GitHub. This requires the `gh` CLI to
be available. This is intended to be used along with an issue template that
sets the structure of a spec suggestion issue.

A new issue template allows to collect suggestions to add a new spec in a
structured issue. As for the command-line, the issue template essentially asks:
"what is the spec URL?", and leaves the possibility to specify a JSON object to
define additional properties. That is not super friendly but should not be
useful in the vast majority of cases. Issue creation triggers a bot that runs
the CLI and reports the result in an issue comment. The same issue comment is
reused if the bot runs multiple times on the issue (e.g., because the issue was
edited).

Once repo maintainers are satisfied with the result, they can add an
`@browser-specs-bot pr` to trigger the creation of PR out of the issue.

NB: The actions require the creation of an `@browser-specs-bot` account and of
a `BOT_TOKEN` secret to run the `gh` CLI command with appropriate rights.

This is a first stab. The code could benefit from refactoring here and there,
and I'm sure that the code/workflow will need to be adjusted.

The README and CONTRIBUTING files were adjusted.

Next steps include:
- Use the CLI to test pull requests (for when people suggest a spec through a
PR).
- Use this mechanism to propose new specs reported by `find-specs.js`
- Add mechanism to say something like `@browser-specs-bot monitor Too early` in
a spec suggestion issue to add the spec to the monitor/ignore lists.