Relax permission constraint on PR submission job (#1275)
Via https://github.com/w3c/browser-specs/issues/1274#issuecomment-2020227300

This moves the permission check from `OWNER` (which is never returned for
organization repos) to `MEMBER` (which is looser, especially in the case of
W3C that has thousands of organization members).

The permissions check is not absolutely needed for now, it just prevents a bit
of spam. The check would become necessary if we decide to skip the PR phase and
have the bot merge the change to the `main` branch directly though.