Test monitor/ignore entries against main list, adjust find-specs
This addresses #238, adding checks that there are no entries in the monitor or
ignore list that are already in the main list.

This also adjusts the find-specs script to:
1. skip fragments in repository homepage URLs (`WICG/close-watcher` does that)
2. also look at spec pages to find URLs, to avoid reporting a spec that is
already in the list (`WICG/close-watcher` also does that as it now points to
an HTML page)
3. compare strings in a case insensitive way, as we sometimes end up with
different casing for repo names and URLs.

A couple of entries needed to be removed from the monitor/ignore list as a
result ;)