Add option to tell the bot to monitor/ignore a spec
With this update, on top of adding the suggested spec to `specs.json`, the bot
may now be instructed to add the suggested spec to the monitor or ignore list.

For instance,  `@browser-specs-bot monitor because the spec is an early draft.`
would instruct the bot to add the spec to the monitor list, with the rationale
"the spec is an early draft."

Under the hoods, the CLI now accepts an issue number for the monitor and ignore
commands, as well as the internal ID of the GitHub issue comment that contains
the command (and thus the rationale).

Command is case-insensitive. The rationale may be provided directly after the
command, or after a `because`, `for`, `as`, `since`. The code expects the
rationale to extend until the end of the comment. For instance, all the
following comments are equivalent:

```
We should wait. @browser-specs-bot monitor the spec is an early draft.
```

```
@browser-specs-bot monitor as the spec is an early draft.
```

```
We should wait.
@browser-specs-bot monitor
Because the spec is an early draft.
```

The monitor and ignore logic are now handled by the same function.

The build logic now also reports an error when it fails to retrieve a URL from
the GitHub issue it received as argument.