Make find-specs create one issue per spec
The `find-specs.js` job used to report all candidate specs at once in an issue.
With this update, it will report candidate specs in dedicated "new spec" issues
that can then be checked and acted upon through bot commands.

The job only creates an issue if one does not already exist. To avoid
degenerative behavior, the job creates 5 issues per run at most. This number
can be adjusted through a command-line option.

The job no longer creates a pull request to add repositories without published
content to the monitor list. In practice, I'm not clear why we were doing that:
if there is no published content, there's nothing super interesting to look at,
monitoring the repository merely acknowledges the fact that we're aware that
the repository exists and could eventually host a real spec. Or are we worried
that we might miss a spec hidden in a repo somewhere? Looking at history, the
only time when we rejected an "Identify new monitored repos" PR was a few
months ago because we decided to add the repository to the ignore list instead,
see #1154.

In the interest of keeping an eye on repositories, the job will continue to
report repositories without published content in the logs, so that we can at
least dig up the history if we realize that we missed a repo at some point, and
wonder why.

Not done here but dropping repo entries from the monitor list that come with a
"no published content yet" comment would greatly reduce the size of the monitor
list without causing any weird consequence, at least as far as I can tell.

I did not see a quick and easy way to reuse the `suggest-spec.yml` issue
template to create the issue. As a result, the issue body is also hardcoded in
`find-specs.js`. We'll need to check the script when we update the template.