Look for auth token in GITHUB_TOKEN (was GH_TOKEN)
The `gh` command that we now use in some of our jobs look for a `GITHUB_TOKEN`
environment variable to find the authentication token, while our code rather
used `GH_TOKEN`.

The job that checks a suggested spec uses both our code and the `gh` command.
To avoid having to set the token to two different environment variables, this
updates rather aligns everything to `GITHUB_TOKEN`.