Make nightly URL of CSS Viewport explicit (#1320)
The build logic expects that, for CSS specs, when a /TR/ spec that has some
level links to a nightly draft without level, then it should add the level
back. That no longer works for CSS Viewport, as css-viewport-1 no longer exists
on the draft CSS server.

This update forces the nightly URL of the CSS Viewport to be without level (and
fixes the build code to actually use that URL when it is specified).