Force series title for orientation-event
Title of the `orientation-event` spec moved to "Device Orientation and Motion".
The W3C API still returns the previous title for the series. The inconsistency
gets caught by the test.

Request to update the W3C API sent. In the meantime, this forces the title of
the spec series to the new one (note the need to set the series `shortname` as
well, only needed because the schema currently requires the property).