Handle FPWD for nightly, fix orientation-event series info, log filename responses (#1330)
* Auto-convert "FPWD" status to "ED" for nightly spec

Build currently chokes on the ED of a CSS spec that incorrectly claims it is a
First Public Working Draft. We already handled a few exceptions to the rule to
accommodate such cases. This adds First Public Working Draft to the list.

* Force series title for orientation-event

Title of the `orientation-event` spec moved to "Device Orientation and Motion".
The W3C API still returns the previous title for the series. The inconsistency
gets caught by the test.

Request to update the W3C API sent. In the meantime, this forces the title of
the spec series to the new one (note the need to set the series `shortname` as
well, only needed because the schema currently requires the property).

* Determine filenames: accept more HTTP status codes, report unexpected ones

The build job currently fails to determine the filenames of /TR versions of W3C
specs. The logs do not contain any useful information for now. This adjusts the
logic to:
1. accept all 2xx codes (just in case there's some network caching involved)
2. report any other code (except 404, which would indeed mean that the resource
does not exist).

That should not fix the issue but will hopefully provide hints about it.