Use actual previous index to determine filename (#1334)
If the code cannot determine the filename of a spec, it reuses the previously
known filename... at least in theory. In practice, the code was not loading the
previous index but the intermediary build step, which does not exist when the
build is run as part of a job.

This makes the code load the previous `index.json` file. Filename is the only
place where that previous index is used.