Force filename to Overview.html for W3C /TR URLs (#1340)
W3C servers have started to become stricter about rate limits, making build
jobs fail with "429 Too Many Requests" responses received. In practice, as
reported in #1333, all /TR specs have an "Overview.html" but two. This update
hardcodes the rule: "All /TR URLs have an Overview.html filename" and adds the
two exceptions to the rule to specs.json.

Note the `release` property was not yet allowed in `specs.json` (because we
never had a need for it). The schemas needed to be updated as a result. To
avoid requiring properties under `release` that are useless in specs.json,
the definition was split into two: one for specs.json without constraint,
and one for index.json that reuses the first one but also adds required
constraints.