Force update of index.json, adjusting discontinued entry
The build code gets lost with the `custom-state-pseudo-class` entry because the
URL now redirects to a section of the HTML spec. Goal is to make the entry
discontinued. Once that is done, the build code should reuse the info from
index.json directly (the discontinued state is a frozen state).

This update forces the standing of the entry to discontinued, reusing previous
info for other entry properties. This should fix the issue at hand and allow
build to resume.

The other updates are triggered by recent changes to `specs.json` that have not
yet been integrated in `index.json` because the build could not succeed. No
manual changes were made to these updates.