Make `multipage` flag specifically target release and/or nightly (#1345)
The job still fails to load the ED of the CSS 2.2 spec to extract the list of
pages, even with optimized loading (added in last commit). Something's rotten
in the State of Network.

In practice, build does not *need* to load the ED of the CSS 2.2 spec. It just
does it because the spec is identified as multipage due to the /TR version. To
avoid unnecessary spec loads, this update converts the `multipage` flag from a
boolean to a string with 3 possible values: "all", "release", "nightly". The
extraction of the list of pages is now only done on the version targeted by the
`multipage` property.

The update also serializes the extraction of pages to avoid sending too many
requests at once. There are only a handful of specs for which we need to look
for pages, serialization should not make build much longer than it already is.