Refresh throttling logic (based on code used in Reffy) (#1360)
This imports the adjustments made to the throttling logic in Reffy. In the end,
it seems useful to adopt the same strategy regarding sleeping between requests:
2s for CSS server, 1s for www.w3.org, and 100ms for other origins.

This raises the question of whether that logic should rather be integrated into
throttled-queue.js if we use the same logic everywhere. Oh well ;)

Refresh also skips CSS resources as done in Reffy.