Make css-overflow-5 a delta spec, add related test (#1389)
The css-overflow-5 spec was added last week as a full spec, but css-overflow-4
is a delta spec. On top of being incorrect because level 5 is also a delta
spec, the succession full > delta > full confuses Reffy, who considers that
level 5 is the "latest version" it should report extracts for, instead of
level 3: it ends up with creating a `css/css-overflow.json` extract for level
5, `css/css-overflow-4.json` for level 4, and no extract for level 3.

This update makes css-overflow-5 a delta spec and adds a test to guarantee that
a delta spec can never be followed by a full spec in its series.