Deal with ARIA exodus
More specs have moved in the past few hours. Migration of the html-aam spec is
still pending but there's an open PR for that, and the spec already exists
under the ARIA mono-repository.

This update hardcodes the rule "all specs whose nightly URL end with -aam or
-aria are in the w3c/aria repository". Except for html-aria, which is developed
by the WebApps WG in a different group. The logic now also finds the source
file on its own in that repository.

This avoids having to add `repository` and `sourcePath` in `specs.json` for all
these specs. The `accname` spec remains an exception to the rule because it
does not end with `-aam` or `-aria`.