Fix specs.json tests when run from the "build" CLI command (#1454)
When the tests are run as part of a "build" command run from the CLI, the tests
on `specs.json` and `index.json` do not actually run on the files but on copies
amended with would-be updates. The path to the copy is read in an environment
variable. The `specs.json` tests were run against the `index.json` copy, likely
due to a hasty copy-and-paste in a previous update. That explains the weird
test failures reported in the latest spec suggestion issues.