Update URL of css-values-5, fix duplicates in diff build
The CSS Values 5 spec was publised as First Public Working Draft last week.

There are 3 specs in the series (Levels 3, 4 and 5). The diff build script
tried to build the other specs in the series twice, one time for the "add"
operation and one time for the "delete" operation, which made tests fail.
The code now adds specs in the same series only once.