Drop digital-identities entry, now digital-credentials (#1501)
The digital-credentials repo was added to the monitor list back in January
(#1174). The digital-identities spec was added shortly afterwards (#1204),
but was using a different repository at the time.

Then digital-identities became digital-credentials, but we missed that.
Instead we recently reviewed the list of monitored specs and decided to
add digital-credentials (#1477).

That means we have twice the same spec in the list. This update drops
the digital-identities entry and adds the name as a former name of
digital-credentials, since that's the name under which the spec is now
progressing.