Fix monitor/ignore bot commands (#1516)
The monitor/ignore bot commands expected to receive as comment ID the same ID
as the one returned by the `gh` CLI. That would have been way too easy! The
comment ID in a workflow is an integer while the ID returned by the `gh` CLI is
a string (probably an encoding of the number?). The code now computes the URL
of the comment from the comment ID and uses that instead to find the comment in
the list returned by `gh`.

The update also fixes a bug in the regular expression used to extract the
actual comment (extracted string included the "because" prefix).

Closes #1496.