Stop fetching info from Specref for WICG specs and W3C ED
This provides a more generic solution to #1606.

Browser-specs has become the main source for Specref for WICG specs and
W3C Editor's Drafts. The code no longer attempts to fetch info for these specs
from Specref as a result.

Tested locally, this will adjust a handful of spec titles (to match the actual
spec title, which seems good). Interestingly, it will also adjust a few WICG
entries for which the spec says the status is "Unofficial Draft", which will
also switch the standing for these specs to "pending". That seems good as well
although we may want to review these changes before we release a new version.
I propose to do that separately in any case. Specs in that category include:
Web Share Target API, preferCurrentTab, the CSS Parser API.

Note: Specref is still being used as source, mainly for WHATWG specs.