Force a good standing for CSS Parser API and preferCurrentTab (#1642)
See discussion in:
https://github.com/w3c/browser-specs/pull/1628#issuecomment-2586337351

Specs used to come with a "good" standing in browser-specs because data came
from Specref. The specs themselves continue to say they are unofficial drafts,
and so build code now considers that these specs should have a "pending"
standing.

That standing seems correct, but WPT already has actual tests for the CSS
Parser API (on top of IDL tests), and already integrated the preferCurrentTab
IDL as well (just a partial dictionary definition in practice). This update
forces the standing to "good" for continuity.