Retrieve WHATWG spec titles from WHATWG database (#1666)
Build code used to rely on Specref to get the title of WHATWG specifications.
This update makes it fetch info for WHATWG specs from the WHATWG database
directly. To save one request, the code leverages the workstreams database,
also used by fetch-groups, instead of the biblio file.

On top of adding a new `whatwg` value to the `"source"` field, this update will
also fix the titles of the WHATWG specs: they end with "Standard" in Specref
but, while that matches the `<title>` tag, the actual spec title in the `<h1>`
and the title in the WHATWG database don't end with "Standard". #docallmeDOM

The update turns the `fetchJSON` function into a utility function. This is
going to save a few requests (not that many!) that are common between
fetch-info and fetch-groups. Specific functions in fetch-info were also
adjusted not to do anything when there are no specs of interest in the list
(this speeds up tests a bit, but has no impact on a full build since, by
definition, there are specs of interest in the full list...)