Stop fetching spec info from Specref (#1668)
This completely drops the logic that used to fetch spec info from Specref.
Motivation for the change is to avoid having a circular dependency between
browser-specs and Specref that makes it cumbersome to make some updates.

In practice, there remained only 6 specs for which browser-specs was leveraging
Specref: 3 from ISO, 2 from TC39 and the Web Bluetooth API.

One consequence is that the title of the TC39 specs will change because the
actual specs include a registered mark and the year, which are now captured,
leading to "ECMAScript® 2025 Language Specification". Maybe not ideal, but then
that's per spec...

README and comments updated in the code to remove mentions of Specref.