Fold CSS 2.1 and CSS 2.2 into a CSS 2 entry
The W3C API was updated to make `CSS2` the actual spec and to have `CSS21`
redirect to `CSS2`. This follows from discussions with @fantasai, @tabatkins
and @deniak to clarify that revision 1 and revision 2 of CSS Level 2 should be
viewed as just CSS Level 2.

This update folds the CSS 2.1 and CSS 2.2 entries into one CSS 2 entry
accordingly:
- Shortname is `CSS2`
- Published version is at https://www.w3.org/TR/CSS2/
- Editor's Draft is at https://drafts.csswg.org/css2/

A side benefit is that there will no longer be duplicated definitions in Webref
for CSS 2.1 and CSS 2.2. That should remove one of the de-duplication rules.

Note the W3C API continues to have two entries because CSS 2.2 got published as
a separate spec in 2016. Hopefully, the specs will eventually get folded there
as well.