Fix lookup of series for diff building (#1714)
When diff building a spec, other specs in the same series need to be re-built
as well. The code correctly picked up the other specs in the same series but
forgot that entries in `specs.json` may either be a string or an object with
a `url` property when it removed duplicates, leading to an incomplete list in
the end.

Fixes #1713