Handle CSS2 series URLs as an exception to the rule (#1730)
For CSS2, we create a "CSS" series in browser-specs different from the "css"
series for CSS snapshots but the W3C API mixes both, and the URL
https://www.w3.org/TR/CSS/ that would logically be computed as series URL for
the CSS series actually returns a CSS Snapshot. The update makes the code use
the CSS2 spec URLs instead.

An alternative would be to create a "CSS2" series instead, but that's well
aligned with the current data in the W3C API.