Further extend support for `shacl12-*` specs
The specs belong to a common repository that uses a slightly different pattern
that wasn't yet supported. Also, as for other specs following the `fooXY-bar`
pattern, the series title does not completely look like the spec's title. The
test now considers the `fooXY-bar` case as an exception to the rule.