Rewrite find-specs, drop monitor/ignore data files
This is a rewrite of the find-specs logic motivated by the idea of gathering
the list of specs that are being monitored, or that have been voluntarily
ignored, from GitHub issues instead of from JSON data files.

One JSON file gets introduced to handle repositories that are known to contain
multiple specifications in a more declarative way.

Also, a hardcoded list of working groups that do not directly standardize
technologies for web browsers appears in compute-categories. That isn't ideal,
but then the list does not need to be perfect either.

The find-specs looks at open issues that have a "new spec" label and does not
report specs that we're already aware of. It also looks at closed issues that
have an "ignore" label, assuming that these specs should not be reported again.

The monitor-specs job was rewritten as well. It no longer updates the now gone
data files and no longer creates a pull request. Instead, it loops through the
open issues that have a "new spec" label and adds a comment and a "review"
label to those that, it detects, target a spec that got updated since the last
time that the issue was reviewed (determined as the last time that the "review"
label got removed).

Interestingly, the whole logic does not need the "monitor" label. That label
may still be useful to highlight the fact that we decided to delay the addition
of a spec to the list to external readers, but it does not affect the logic:
any open issue with a "new spec" label gets monitored, and a "review" label is
added when an update was detected.