Support ED to FPWD transition (#1769)
When it finds a candidate /TR URL for a spec that was already in the list
through its ED URL, the find-specs script now also adds a link to the ED in
the rationale section of the issue.

When it finds a candidate /TR URL for a spec for which there is a pending issue
for the ED URL, the find-specs script now creates a new issue for the /TR URL
that links to the ED URL issue in the rationale section. The script also closes
the ED URL issue with a "Superseded by #xxxx" comment.

The build logic leverages the ED link to also propose the deletion of the
ED entry from the list at the same time as the /TR URL is added.

The update also introduces improvements to the comparison logic in the
find-specs script, and skips the check on open issues before creating a new one
(that's no longer needed since, by definition, we look at open issues before we
suggest candidate specs).