Crawl new spec proposals with Reffy and report
When a spec gets proposed through a GitHub issue, browser-specs now runs Reffy
against it and leverages the new `summary` option in Reffy to get a report in
Markdown that it can integrate in the comment it writes. This gives us more
clues about what adding a spec would change in Webref.