Crawl new spec proposals with Reffy and report (#1787)
When a spec gets proposed through a GitHub issue, browser-specs now runs Reffy against it and leverages the new summary option to get a report in Markdown that it can integrate in the comment it writes. This should give us more clues about what adding a spec would change in Webref.

The update also fixes a minor code hiccup that made the job crash on the recently created spec proposal issues.

A manual job trigger was added to give a way to run the job again on an issue (e.g., because the spec got updated since the last time the job ran and we'd like a new report, or because the job failed for some transient reason).

Note: To be monitored, but we may need to truncate the Markdown summary that Reffy produces, because GitHub restricts issue comments to something like 65536 characters (*4 if the report only contains ASCII characters, I believe). An obvious way to do that without losing too much is to drop all links to the xref search fro mthe report as they consume a large number of characters.