Update URL of Subresource Integrity (#1805)
There used to be one Recommendation for Subresource Integrity, whose shortname
was `SRI`. A new level 2 was published as First Public Working Draft today.
Shortnames got updated accordingly: `sri-2` for the new level, and `sri-1` for
the previous Recommendation, and `SRI` now returns that new level. Our build
code detects the redirection and does not like it.

This updates the URL in the list to use `sri-2` (recording `SRI` as a former
name). We could keep an entry for `sri-1` but we usually only do that when
the current spec is not the latest level.