Throttle JSON fetches per origin
The `fetchJSON` function was only applying loose throttling, but
https://raw.githubusercontent.com is starting to return 429 responses after a
while (we send a few hundreds of requests to retrieve `w3c.json` files).

Documentation does not say much, but it would seem the limit is around ~5000
requests per hour. This update makes the function sleep 1000ms in between
requests to a given origin.

This will likely add ~10mn of (idle) build time.