Fetch and use info from validate-repo when possible
The validate-repos project collects all known `w3c.json` file, and contains
the information we need to map repo names to groups for a good chunk of specs.

This still does not work well for /TR specs because we guess the name of the
repository from the spec's shortname, but that's still better than what we used
to have!