Leverage validate-repos and throttle requests on w3c.json files (#1831)
* Throttle JSON fetches per origin

The `fetchJSON` function was only applying loose throttling, but
https://raw.githubusercontent.com is starting to return 429 responses after a
while (we send a few hundreds of requests to retrieve `w3c.json` files).

Documentation does not say much, but it would seem the limit is around ~5000
requests per hour. This update makes the function sleep 1000ms in between
requests to a given origin.

This will likely add ~10mn of (idle) build time.

* Fetch and use info from validate-repo when possible

The validate-repos project collects all known `w3c.json` file, and contains
the information we need to map repo names to groups for a good chunk of specs.

This still does not work well for /TR specs because we guess the name of the
repository from the spec's shortname, but that's still better than what we used
to have!