Add new RDF/SPARQL WG to list of "non-browser" groups
The group used to be known as the "RDF-star" working group. New name needed to
be added to the list of non-browser groups to avoid categorizing the group's
deliverables as "browser" specs.

Note: the list still contains the previous name in case some data remains
that still references that name.