Drop dependency on Mocha in CLI (#1900)
This replaces the CLI logic that was based on Mocha for running tests against
the index that a proposed change would create, with logic based on Node's test
module.

A simple test on the CLI was also added to make sure that it's at least possible to
run it, so that we don't miss problems next time.

Fixes #1899.