Use ISO Open Data datasets to process ISO specs
Logic to process specs from ISO was to "browse" the ISO web site to extract
the spec title and group name. That approach fails more and more, essentially
because bots are not welcome on the web site.

This update drops custom logic to browse the ISO web site and leverages
machine-readable datasets provided through the ISO Open data project instead:
https://www.iso.org/open-data.html

Now, the thing is, browser-specs only contains 3 specs from ISO for now, and
their information is unlikely ever going to change. And the datasets are huge
(~70MB for the main catalog). To avoid wasting resources for nothing, a new
build command gets introduced:

```
npm run build-skip-iso
```

The command skips fetching the ISO catalog and reuses previous info about the
ISO specs.

That new command is used for all automated builds from Monday to Friday. On
Saturday and Sunday, a full build gets made, only once per day (running
multiple builds during the week-end is not very useful in any case as we're
unlikely ever going to act on them).

Under the hoods, the new command uses a `--skip-fetch` CLI option, which could
perhaps be extended afterwards to skip additional network fetches.

The `source` of ISO entries changes from `spec` to `iso`.